Updated with PR comments

+++ src/backend/pagos/app/models/reembolso.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, ForeignKey
+from sqlalchemy import Column, String, Float, ForeignKey
 from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.postgresql import UUID
 from travelhub_common.models import BaseModel
@@ -7,7 +7,7 @@ from app.models.pago import Pago
 class Reembolso(BaseModel):
     __tablename__ = "reembolsos"
 
-    monto = Column(Integer, nullable=False, default=0)
+    monto = Column(Float, nullable=False, default=0)
     razon = Column(String, nullable=True)
     estado = Column(String(50), nullable=False)
